8.14 bugfix

+++ src/controllers/videoController.js
@@ -17,6 +17,7 @@ export const watch = async (req, res) => {
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
+
   return res.render("video/watch", { pageTitle: video.title, video });
 };
 
@@ -25,6 +26,9 @@ export const getEdit = async (req, res) => {
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
   }
   return res.render("video/edit", {
     pageTitle: `Edit "${video.title}"`,
@@ -35,9 +39,13 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  const isVideo = await Video.exists({ _id: id });
-  if (!isVideo) {
+  const video = await Video.findById(id);
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
   }
 
   await Video.findByIdAndUpdate(id, {
@@ -82,6 +90,13 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
